# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Projetlvre",
        "description": "une sorte de biblio en ligne",
        "version": "1.0"
    },
    "paths": {
...
        "/api/paiement": {
            "get": {
                "tags": [
                    "Payment"
                ],
                "summary": "Retrieve payment details including total cart price",
                "description": "Get payment information",
                "operationId": "cbd6b41b431b3a1368a1462ce2a70cc0",
                "responses": {
                    "200": {
                        "description": "Payment information retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "type": "object"
                                        },
                                        "amount": {
                                            "type": "number",
                                            "example": 75.5
                                        },
                                        "msg": {
                                            "type": "string",
                                            "example": "Payment information retrieved successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/validatepaiement": {
            "post": {
                "tags": [
                    "Payment"
                ],
                "summary": "Validate payment and create an order",
                "description": "Validate payment",
                "operationId": "c56491a7b05c5b20b538d3fb09405a3f",
                "responses": {
                    "200": {
                        "description": "Payment completed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "Purchase completed successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Insufficient balance",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "Insufficient balance to complete the purchase"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No item in cart",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "Error, No data in your cart"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get authenticated user data",
                "description": "Retrieve information about the authenticated user. If the user is a seller, the API returns their books. If the user is a buyer, it returns all available books.",
                "operationId": "fd8e4e27fd1a9e98a6638cec7e2126ca",
                "responses": {
                    "200": {
                        "description": "User data retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "description": "Authenticated user information",
                                            "type": "object"
                                        },
                                        "userlivre": {
                                            "description": "Books belonging to the seller",
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "allbooks": {
                                            "description": "All books available for buyers",
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "msg": {
                                            "type": "string",
                                            "example": "User books retrieved successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/rechargesolde": {
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Recharge user account balance",
                "description": "Add funds to the authenticated user's balance.",
                "operationId": "7994a17655a98ad300d34b48a832f158",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "solde"
                                ],
                                "properties": {
                                    "solde": {
                                        "description": "Amount to add to the user balance",
                                        "type": "number",
                                        "example": 100
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Balance successfully recharged",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "Balance successfully recharged"
                                        },
                                        "user": {
                                            "description": "Updated user information",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}